requirements.txt
google-generativeai==0.8.5
streamlit==1.47.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.3.1
faiss-cpu==1.11.0.post1
sentence-transformers==5.0.0
python-dotenv==1.1.1
pypdf==5.8.0
torch==2.7.1
transformers==4.53.3
